Removed redundant code from store that replicates code in phyx.js.
